GPT/requirements.txt
streamlit
langchain_google_genai.chat_models
dotenv

